main controller

+++ src/main/java/e2p/icotp/model/Loan.java
@@ -21,14 +21,14 @@ public class Loan {
     private DoubleProperty principal;
     private DoubleProperty interest;
     private DoubleProperty penalty;
-    private DoubleProperty due;
+    private LongProperty due;
     private DoubleProperty paid;
     private DoubleProperty balance;
     private StringProperty status;
 
     public Loan(int loan_id, Loaner loaner_id, LocalDate release_date, int term, LocalDate maturity_date,
             double principal,
-            double interest, double penalty, double due, double paid, double balance, String status) {
+            double interest, double penalty, long due, double paid, double balance, String status) {
 
         this.loan_id = new SimpleIntegerProperty(loan_id);
         this.loaner_id = new SimpleObjectProperty<>(loaner_id);
@@ -38,7 +38,7 @@ public class Loan {
         this.principal = new SimpleDoubleProperty(principal);
         this.interest = new SimpleDoubleProperty(interest);
         this.penalty = new SimpleDoubleProperty(penalty);
-        this.due = new SimpleDoubleProperty(due);
+        this.due = new SimpleLongProperty(due);
         this.paid = new SimpleDoubleProperty(paid);
         this.balance = new SimpleDoubleProperty(balance);
         this.status = new SimpleStringProperty(status);
@@ -81,7 +81,7 @@ public class Loan {
         this.penalty.set(penalty);
     }
 
-    public void setDue(double due) {
+    public void setDue(long due) {
         this.due.set(due);
     }
 
@@ -130,7 +130,7 @@ public class Loan {
         return this.penalty.get();
     }
 
-    public Double getDue() {
+    public Long getDue() {
         return this.due.get();
     }
 
@@ -179,7 +179,7 @@ public class Loan {
         return this.penalty;
     }
 
-    public DoubleProperty getDueProperty() {
+    public LongProperty getDueProperty() {
         return this.due;
     }
